fix: Energiezähler robust (double, Wrap-/Glitch-/Ausfall-Schutz)
- totalEnergy_Wh/lastEnergy_Wh von float auf double: die float-Mantisse (24 Bit)
  schluckt bei hohen kWh-Summen kleine 5-s-Inkremente -> Zaehler fror ein.
  Flash-Persistenz auf Version 2 (writeDouble); Version 1 (float) wird beim
  Einlesen migriert.
- INA228: getEnergy()==0 (fehlgeschlagener read40) wird als Sample verworfen
  statt den gesamten HW-Akkumulator vom Zaehler abzuziehen; ein negativer Delta
  (40-bit-Register-Ueberlauf) setzt nur die Baseline neu, statt den Zaehler zu
  korrumpieren.
- INA226: nur plausible Zeitscheiben (<30 s) integrieren und lastUpdateTime bei
  (Neu-)Verbindung neu ansetzen -> kein Fabrizieren von Energie ueber die
  Dauer eines I2C-Ausfalls.

+++ src/MeasuringModule.h
@@ -84,8 +84,8 @@ private:
     float busVoltage_V = 0.0;
     float current_A = 0.0;
     float power_W = 0.0;
-    float totalEnergy_Wh = 0.0;
-    float lastEnergy_Wh = 0.0;
+    double totalEnergy_Wh = 0.0; // double: float-Mantisse friert bei hohen kWh-Summen ein
+    double lastEnergy_Wh = 0.0;
     
     float temperatur_C = 0.0;
